v1.2.0
Added additional onChange 2 parameter compatibility version and added ability to specify initial setting (and added documentation to match the new (current) implementations).  Moved threading functions into static Compatibility functions so that we can reference in case we're in a class that shadows the same function name (like running background {} from within a view that is trying to create a view).  Added returning background { } calls for cases where we need to await the results of the long-running background task.  Re-worked debugLevel features of debug statements so we aren't switching threads with the print statement to ensure debug statements output immediately and don't get printed out of order.  Added Compatibility.isDebug flag for testing if we've built for release or debug.  Added additional Backport code including `scrollClipDisabled()`.  Added set additions for OrderedSet and OrderedDictionary and added merging/interoperability between OrderedDictionary and Dictionary.