v1.11.9
Added import for Color to the RadialLayout preview so that included packages won't warn.  Change MixedTypeFieldDictionary to default to an unordered dictionary so JSON encoding matches and cleaned up and simplified and optimized code.  This will mean that the output is inconsistent but that’s what the `sortedKeys` option is for.  Added Backport for `controlSize`.  Re-worked all the `swiftUIValue` conversions to be consistent.  Re-worked `firstKey` extension to work on any DictionaryConvertible so that it works on ordered dictionaries as well.